requirements.txt
asgiref==3.4.1
astroid==2.8.2
coverage==6.0.1
dj-database-url==0.5.0
Django==3.2.8
django-otp==1.1.1
djangorestframework==3.12.4
flake8==3.9.2
gunicorn==20.1.0
isort==5.9.3
lazy-object-proxy==1.6.0
mccabe==0.6.1
platformdirs==2.4.0
psycopg2-binary==2.9.1
pycodestyle==2.7.0
pyflakes==2.3.1
pylint==2.11.1
pylint-django==2.4.4
pylint-plugin-utils==0.6
python-decouple==3.5
pytz==2021.3
qrcode==7.3.1
sqlparse==0.4.2
toml==0.10.2
typing-extensions==3.10.0.2
whitenoise==5.3.0
wrapt==1.12.1
